Refactor category management into dedicated module
